+++ src/Symfony/Bundle/SecurityBundle/Tests/Functional/app/PasswordEncode/bundles.php
@@ -12,4 +12,5 @@
 return [
     new Symfony\Bundle\SecurityBundle\SecurityBundle(),
     new Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
+    new Symfony\Bundle\SecurityBundle\Tests\Functional\Bundle\TestBundle(),
 ];